some fixes and add module tests

+++ src/main/java/com/MomsDeveloper/CSVExporter.java
@@ -4,7 +4,7 @@ import java.io.FileWriter;
 import java.io.IOException;
 
 public class CSVExporter {
-    public static void exportToCSV(double start, double end, double step, double terms, String filename, SeriesFunction func) {
+    public static void exportToCSV(double start, double end, double step, int terms, String filename, SeriesFunction func) {
         try (FileWriter writer = new FileWriter(filename)) {
             writer.append("X, Result\n");
             for (double x = start; x <= end; x += step) {
