+++ src/main/java/com/MomsDeveloper/FunctionSystem.java
@@ -28,7 +28,7 @@ public class FunctionSystem extends BaseFunction {
         this.log2 = log2;
     }
 
-    public double calculate(double x, double terms) {
+    public double calculate(double x, int terms) {
         if (!checkParams(x, terms))
             return Double.NaN;
         if (x <= 0) {
