+++ src/main/java/com/MomsDeveloper/SeriesFunction.java
@@ -1,5 +1,5 @@
 package com.MomsDeveloper;
 
 public interface SeriesFunction {
-    public double calculate(double x, double terms);
+    public double calculate(double x, int terms);
 }
