+++ src/main/java/com/MomsDeveloper/trigonometric/Cot.java
@@ -9,7 +9,7 @@ public class Cot extends BaseFunction {
         this.sin = sin;
         this.cos = cos;
     }
-    public double calculate(double x, double terms) {
+    public double calculate(double x, int terms) {
         if (!checkParams(x, terms)) return Double.NaN;
         double sin_res = sin.calculate(x, terms);
         double cos_res = cos.calculate(x, terms);
